mathai_backend/requirements.txt
# Testing
pytest>=8.0.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Authentication
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
email-validator>=2.0.0

# HTTP Client
requests>=2.31.0
httpx>=0.26.0

# Math & Computation
sympy>=1.12

# Database
sqlalchemy>=2.0.0
alembic>=1.17.0
aiosqlite>=0.21.0

# Configuration
pydantic-settings>=2.1.0

# Performance
uvloop>=0.19.0
httptools>=0.6.1

# Rate Limiting
slowapi>=0.1.9

# File Locking
filelock>=3.13.0

# Metrics & Monitoring
prometheus-fastapi-instrumentator>=6.1.0

# Logging (optional structured logging)
python-json-logger>=2.0.7
